Introduce the ES client and constants modules
The current commit introduces the Elasticsearch client and the constants
modules.

The `Client` module will be used to define all the Elasticsearch related
functions. Currently, we introduce a first (dummy) version of the
`wait_for_status` method, that will be used to wait for the desired
cluster health status.

Finally, we add a `Constants` module, to include all the tool specific
constants.