Add functionality to toggle between shard allocation settings
This patch introduces all the necessary functionality to toggle the
transient shard allocation settings of an Elasticsearch cluster.

In more details, we first introduce the `Utils.hs` module containg a new
method, the `curlPutString`, which is used to perform PUT requests and
return the request's return code and the respective response as a Haskell
String. It is a method that is not defined in the `Network.Curl` module.

Then we introduce the `Json.hs` module that currently defines the ES
transient cluster settings type. That module will be used to define
all future JSON/dict types.

Finally, we add the actual method to switch between shard allocation
settings, named `shardAllocToggle`.