Modify the package directory structure
The current commit creates a basic directory hierarchy for our package.
It separates the source and executable files of our package in order to
be easier to package it at a later point.

In more details, the source files are moved under the
`src/Elastic/RollingRestart` location and any executables under the
`executables` directory. Also, note that the utility related modules
are moved under a new directory, named `Utils`.